Prevent dealiasing for type aliases defined outside scala and Predef
The current formatting scheme is inconsistent when showing types of values of
type aliases. For example, given `trait State[S, A]` and `type IntState[A] =
State[Int, A]`, the type of `val s: IntState[Int]` is formatted as `State[Int]`.

The problem is that `typeSymbol` (which dealiases the type) is used for
displaying the name of the type, while `typeArgs` (which does not dealias) is
used for formatting the type parameters.

This commit makes the scheme more consistent:
- When types need to be dealiased, `tpe.typeSymbol` is used for the name and
`tpe.dealias.typeArgs` is used for the type parameters
- When types do not need to be dealiased, `tpe.typeSymbolDirect` and
`tpe.typeArgs` are used

The rules for dealiasing (hardcoded in the `shouldDealias` function) are as
such:
- Type aliases defined in the `scala` package are dealiased
- Type aliases that have `Predef` in their prefix are dealiased (to support the
standard `scala.Predef` and user-supplied predefs)
- All other type aliases are not dealiased

Resolves #1858.